post id Long 타입 변환 post 리스트 완료
버튼 + 검색바 + 페이징 css 예정

+++ src/main/java/com/example/dayte/reply/domain/ContentReply.java
@@ -21,7 +21,7 @@ public class ContentReply {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int num; // 댓글 번호
+    private Long num; // 댓글 번호
 
     @JoinColumn(name = "user")
     @ManyToOne(fetch = FetchType.EAGER)
